Extract the websocket server out of API.Server
This also merges withAPIServer and runAPIServer together as most
complexity comes from arguments being passed in.
